Media: Split uploading files out from adding new media
